Fix ImageSequenceReference cloning for int64_t
- The 'Writer' part of SerializableObject needed the same tightening as
  the RapidJSON dispatcher
- this is used for things that serialize and compare JSON (like cloning
  and the `is_equivalent_to` method.
- Because the ImageSequenceReference schema has `int` in it (which
  should at some point be changed to int64_t), it was tripping this
  check on certain platforms

+++ src/opentimelineio/serializableObject.h
@@ -319,7 +319,7 @@ public:
         static bool write_root(any const& value, class Encoder& encoder, ErrorStatus* error_status);
 
         void write(std::string const& key, bool value);
-        void write(std::string const& key, int value);
+        void write(std::string const& key, int64_t value);
         void write(std::string const& key, double value);
         void write(std::string const& key, std::string const& value);
         void write(std::string const& key, RationalTime value);
